Moved studyright parsing to data_mapper, added mapper to api get method, implemented updating of values for studyrights found in db.
